State design pattern added to the interest service

+++ src/main/java/InterestService/InterestsMechanism.java
@@ -3,5 +3,5 @@ package InterestService;
 import java.util.Date;
 
 public interface InterestsMechanism {
-    double generateInterest(int amountOfMoney, double percentage, int howManyCapitalisation);
+    double generateInterest(long amountOfMoney, double percentage, int howManyCapitalisation);
 }
